Update DhtBase.cs
Fix Bug:

Scenario: Connect the sensor (DHT) to the ESP32 and read data from it. Then, unplug the sensor (DHT) from the ESP32. Now, the "ReadThroughOneWire" method returns the last data, and "IsLastReadSuccessful" is true (in other word we can detect unplugging had happened). In this change, "_readBuff" will be cleared, and "IsLastReadSuccessful" will be false if we don’t receive valid data